Fix Pagination Logic

+++ MDFlickerImages/MDFlickerImages/Modal/Network Modal/MDNetworkService.swift
@@ -12,7 +12,7 @@ import MMNetworkManager
 final class MDNetworkService {
     private let urlString: String = "https://api.flickr.com/services/rest/?method=flickr.photos.search&api_key=2932ade8b209152a7cbb49b631c4f9b6&%20format=json&nojsoncallback=1&safe_search=1&per_page=21&text="
     
-    func getFlickerImages(for searchQuery: String, of page: Int, completion: @escaping (Result<FlickerPhoto?, NetworkError>) -> Void) {
+    func getFlickerImages(for searchQuery: String, of page: Double, completion: @escaping (Result<FlickerPhoto?, NetworkError>) -> Void) {
         guard searchQuery.isNotEmpty else {
             completion(.failure(.noSearchQuery))
             return }
